2.0 Release, Full rework of the current system
Me and my team at better vanilla have decided that instead of "Improving" the game, via 3d models, overhauls and drastic changes; We'd shift our focus back to our original goal. A simple streamlined pack that improves the game without making it overbearing, no 3d models no drastic changes. And as such, we have added tons of smaller things that are to many to go over, the eventual hope is that you wont notice much different from the normal game. A very small list of what we've changed are.

>General Sound overhauls

>watermelon and pumpkin stems now match the color of the fruit they produce.

>Better wool textures and upgrades

>Darker dark oak

>Horizontal item nuggets gold iron etc

>Brown leather and all of the associated items.

>Banners on shields now display properly

>Arrows now have a black tip to represent the flint that's required to craft them.

>Polished granite and andesite now show as bricks.

>Water and leaves now sway in the wind.

>The sun and moon are now round.

>Snow has had the blue tint removed from it

>Some locks now have variation.

>Some blocks now have connected textures.

>The crosshair has been changed to a dot.

>Dark UI

>Various small fixes and tweaks (Such as not eating the wood from soup. or fixing the bottom of blaze's)

Thank you for your understanding and support in our redesign if you have any suggestions be sure to put them into the issues section of the GitHub there will be a feature request template. For anything else contact me on discord Xanthus58#3862